sort min_imports alphabetically
Since there is a call to `qsort` it seems like sorting is intended here.
However, currently `qsort` is called on `Name`, but apparently this does
not do the correct thing, since when using this in mathlib, I observed
that the output was not sorted. We fix this by converting to string
earlier and using the `(· < ·)` function to compare the strings.